requirements.txt
# Core dependencies
flask>=2.0.0
aiohttp>=1.0.0
websockets>=10.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# PostgreSQL driver
psycopg2-binary>=2.9.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Code quality (development)
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.0.0
types-requests>=2.28.0
pre-commit>=3.0.0